Delay: Fix Gate rounding to whole issue

+++ Apollo/Devices/Delay.cs
@@ -81,7 +81,7 @@ namespace Apollo.Devices {
 
             Dictionary<string, object> data = JsonConvert.DeserializeObject<Dictionary<string, object>>(json["data"].ToString());
 
-            return new Delay(Convert.ToBoolean(data["mode"]), new Length(Convert.ToDecimal(data["length"])), Convert.ToInt32(data["time"]), Convert.ToInt32(data["gate"]));
+            return new Delay(Convert.ToBoolean(data["mode"]), new Length(Convert.ToDecimal(data["length"])), Convert.ToInt32(data["time"]), Convert.ToDecimal(data["gate"]));
         }
 
         public override string EncodeSpecific() {
